+++ mod-audit-server/src/test/java/org/folio/rest/impl/InventoryAuditApiTest.java
@@ -3,6 +3,7 @@ package org.folio.rest.impl;
 import static io.restassured.RestAssured.given;
 import static org.folio.services.configuration.Setting.INVENTORY_RECORDS_PAGE_SIZE;
 import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
 
 import io.restassured.http.Header;
@@ -46,7 +47,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 public class InventoryAuditApiTest extends ApiTestBase {
 
   private static final String TENANT_ID = "modaudittest";
-  private static final String ENTITY_ID = UUID.randomUUID().toString();
   private static final Header TENANT_HEADER = new Header("X-Okapi-Tenant", TENANT_ID);
   private static final Header PERMS_HEADER = new Header("X-Okapi-Permissions", "audit.all");
   private static final Header CONTENT_TYPE_HEADER = new Header("Content-Type", "application/json");
@@ -81,13 +81,14 @@ public class InventoryAuditApiTest extends ApiTestBase {
   @MethodSource("provideResourceTypeAndPath")
   void shouldReturnInventoryEventsOnGetByEntityId(InventoryResourceType resourceType, String apiPath) {
     var dao = resourceToDaoMap.get(resourceType);
+    var entityId = UUID.randomUUID().toString();
     var changeRecordDto = new ChangeRecordDto();
     changeRecordDto.setFieldChanges(List.of(new FieldChangeDto(ChangeType.MODIFIED, "id", "id", "old", "new")));
 
     var inventoryAuditEntity = new InventoryAuditEntity(
       UUID.randomUUID(),
       Timestamp.from(Instant.now()),
-      UUID.fromString(ENTITY_ID),
+      UUID.fromString(entityId),
       "CREATE",
       UUID.randomUUID(),
       changeRecordDto
@@ -96,10 +97,11 @@ public class InventoryAuditApiTest extends ApiTestBase {
     dao.save(inventoryAuditEntity, TENANT_ID).toCompletionStage().toCompletableFuture().get();
 
     given().headers(HEADERS)
-      .get(apiPath + ENTITY_ID)
+      .get(apiPath + entityId)
       .then().log().all()
       .statusCode(HttpStatus.HTTP_OK.toInt())
-      .body(containsString(ENTITY_ID));
+      .body(containsString(entityId))
+      .body("totalRecords", equalTo(1));
   }
 
   @SneakyThrows
@@ -116,6 +118,8 @@ public class InventoryAuditApiTest extends ApiTestBase {
       .groupId(INVENTORY_RECORDS_PAGE_SIZE.getGroup().getId())
       .updatedDate(LocalDateTime.now())
       .build();
+
+    var entityId = UUID.randomUUID().toString();
     settingDao.update(settingEntity.getId(), settingEntity, TENANT_ID).toCompletionStage().toCompletableFuture().get();
 
     // Create InventoryAuditEntity entities with a year date difference
@@ -126,7 +130,7 @@ public class InventoryAuditApiTest extends ApiTestBase {
       var inventoryAuditEntity = new InventoryAuditEntity(
         UUID.randomUUID(),
         Timestamp.from(Instant.now().plusSeconds(2628000L * i)), // 1 month difference
-        UUID.fromString(ENTITY_ID),
+        UUID.fromString(entityId),
         "CREATE",
         UUID.randomUUID(),
         changeRecordDto
@@ -137,11 +141,12 @@ public class InventoryAuditApiTest extends ApiTestBase {
 
     // Query the API once
     var firstResponse = given().headers(HEADERS)
-      .get(apiPath + ENTITY_ID)
+      .get(apiPath + entityId)
       .then().log().all()
       .statusCode(HttpStatus.HTTP_OK.toInt())
       .assertThat()
       .body("inventoryAuditItems", hasSize(3))
+      .body("totalRecords", equalTo(5))
       .extract().response();
 
     // Extract the last date from the first response
@@ -149,10 +154,11 @@ public class InventoryAuditApiTest extends ApiTestBase {
 
     // Query the API again, passing the last date from the previous response
     given().headers(HEADERS)
-      .get(apiPath + ENTITY_ID + "?eventTs=" + lastDate)
+      .get(apiPath + entityId + "?eventTs=" + lastDate)
       .then().log().all()
       .statusCode(HttpStatus.HTTP_OK.toInt())
-      .body("inventoryAuditItems", hasSize(2));
+      .body("inventoryAuditItems", hasSize(2))
+      .body("totalRecords", equalTo(5));
   }
 
   @ParameterizedTest
@@ -162,7 +168,8 @@ public class InventoryAuditApiTest extends ApiTestBase {
       .get(apiPath + UUID.randomUUID())
       .then().log().all()
       .statusCode(HttpStatus.HTTP_OK.toInt())
-      .body("inventoryAuditItems", hasSize(0));
+      .body("inventoryAuditItems", hasSize(0))
+      .body("totalRecords", equalTo(0));
   }
 
   @ParameterizedTest